+++ soh/soh/Enhancements/randomizer/ShuffleIcicles.cpp
@@ -49,7 +49,7 @@ extern "C" void DrawItemHalo(Actor* icicleActor) {
     // Rotate and draw halo with CMC colors
     Matrix_Translate(icicleActor->world.pos.x + xOffset, icicleActor->world.pos.y + yOffset,
                      icicleActor->world.pos.z + zOffset, MTXMODE_NEW);
-    Matrix_RotateZ(-M_PI / 2, MTXMODE_APPLY);
+    Matrix_RotateZ(static_cast<f32>(-M_PI / 2), MTXMODE_APPLY);
     Matrix_Scale(0.01f, 0.01f, 0.01f, MTXMODE_APPLY);
     OPEN_DISPS(gPlayState->state.gfxCtx);
     gSPMatrix(POLY_OPA_DISP++, MATRIX_NEWMTX(gPlayState->state.gfxCtx), G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_MODELVIEW);
