+++ soh/soh/resource/importer/CollisionHeaderFactory.cpp
@@ -232,12 +232,12 @@ ResourceFactoryXMLCollisionHeaderV0::ReadResource(std::shared_ptr<Ship::File> fi
         }
     }
 
-    collisionHeader->collisionHeaderData.numVertices = collisionHeader->vertices.size();
-    collisionHeader->collisionHeaderData.numPolygons = collisionHeader->polygons.size();
-    collisionHeader->surfaceTypesCount = collisionHeader->surfaceTypes.size();
-    collisionHeader->camDataCount = collisionHeader->camData.size();
-    collisionHeader->camPosCount = collisionHeader->camPosData.size();
-    collisionHeader->collisionHeaderData.numWaterBoxes = collisionHeader->waterBoxes.size();
+    collisionHeader->collisionHeaderData.numVertices = static_cast<u16>(collisionHeader->vertices.size());
+    collisionHeader->collisionHeaderData.numPolygons = static_cast<u16>(collisionHeader->polygons.size());
+    collisionHeader->surfaceTypesCount = static_cast<uint32_t>(collisionHeader->surfaceTypes.size());
+    collisionHeader->camDataCount = static_cast<uint32_t>(collisionHeader->camData.size());
+    collisionHeader->camPosCount = static_cast<int32_t>(collisionHeader->camPosData.size());
+    collisionHeader->collisionHeaderData.numWaterBoxes = static_cast<u16>(collisionHeader->waterBoxes.size());
 
     collisionHeader->collisionHeaderData.vtxList = collisionHeader->vertices.data();
     collisionHeader->collisionHeaderData.polyList = collisionHeader->polygons.data();
